+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Fix overdue search not searching on query changes
 - Bump dagger to v2.43.1
 - Bump fragment to v1.5.1
+- Show chip input text view in overdue search when overdue v2 search is enabled
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342